* adding support for Thermoco temp sensor * allowing long numbers as keys in flag fields (bugfix)

+++ src/main/java/org/sputnikdev/bluetooth/gattparser/spec/Bit.java
@@ -57,12 +57,12 @@ public class Bit {
         return enumerations;
     }
 
-    public String getFlag(byte value) {
+    public String getFlag(BigInteger value) {
         if (enumerations == null) {
             return null;
         }
         for (Enumeration enumeration : enumerations.getEnumerations()) {
-            if (enumeration.getKey().equals(BigInteger.valueOf(value))) {
+            if (enumeration.getKey().equals(value)) {
                 return enumeration.getRequires();
             }
         }
